Added new natives
Added a new native input function - read(message) - message is the prompt that's displayed to the user.
Added dynamic arrays which have the following functions:
set(i, value) - sets the value at the index i, if i is nil then appends the value to the end.
get(i) - returns the value at index i.
length() - returns length of array.

+++ src/pyLox/scanner.py
@@ -1,5 +1,4 @@
 import sys
-from decimal import Decimal
 from token import Token
 from token_type import TokenType
 from error_handler import ErrorHandler
@@ -146,7 +145,7 @@ class Scanner:
             self.advance()
             while self.is_digit(self.peek()):
                 self.advance()
-        self.add_token(TokenType.NUMBER, Decimal(self.source[self.start:self.current]))
+        self.add_token(TokenType.NUMBER, int(self.source[self.start:self.current]))
 
     def peek_next(self) -> chr:
         if self.current + 1 >= len(self.source):
